feat: Add Schedule types
